+++ src/icemac/install/addressbook/__init__.py
@@ -1,1 +1,2 @@
 CURRENT_NAME = 'current'
+DIRNAME_TEMPLATE = 'icemac.addressbook-{}'