+++ src/icemac/install/addressbook/install.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from . import CURRENT_NAME
+from . import CURRENT_NAME, DIRNAME_TEMPLATE
 from .cmd import call_cmd
 import archive
 import argparse
@@ -86,6 +86,20 @@ def symlink(dir_name):
     if os.path.lexists(CURRENT_NAME):
         os.unlink(CURRENT_NAME)
     os.symlink(dir_name, CURRENT_NAME)
+
+
+def make_current(args=None):
+    """Entry point for `bin/make-current-addressbook`."""
+    parser = argparse.ArgumentParser(
+        description='Update the `current` symlink to a given address book '
+                    'version.')
+    parser.add_argument(
+        'version',
+        help='Version number of the icemac.addressbook package you want to '
+             'make the current one.')
+
+    args = parser.parse_args(args)
+    symlink(DIRNAME_TEMPLATE.format(args.version))
 
 
 def main(args=None):